+++ src/java/postback/hod/UnaccountStudent.java
@@ -6,6 +6,7 @@
 package postback.hod;
 
 import entities.Attendance;
+import entities.Department;
 import entities.Student;
 import java.io.PrintWriter;
 import javax.servlet.annotation.WebServlet;
@@ -22,29 +23,37 @@ import utility.PostBackController;
  */
 @WebServlet(urlPatterns = "/teacher/hod/studnets/unaccountstudent")
 public class UnaccountStudent extends PostBackController {
-
+    
     @Override
     public void process(HttpServletRequest req, HttpServletResponse resp, Session session, HttpSession httpSession, PrintWriter out) throws Exception {
         int studentId = Integer.parseInt(req.getParameter("studentId"));
-
+        
+        Department department = (Department) httpSession.getAttribute("department");
+        department = (Department) session.get(Department.class, department.getId());
+        
         Student student = (Student) session.get(Student.class, studentId);
-        if (!student.isVerified()) {
-            student.getClassRoom().getStudents().remove(student);
-            student.setClassRoom(null);
-            student.getSubjects().stream()
-                    .forEach(e -> e.getStudents().remove(student));
-            student.getSubjects().clear();
-            student.getAttendance().stream()
-                    .forEach(e -> e.setStudent(null));
-            student.getAttendance().clear();
-            session.createCriteria(Attendance.class)
-                    .add(Restrictions.isNull("student"))
-                    .list()
-                    .stream()
-                    .forEach(e -> session.delete(e));
-            student.unaccount();
+        
+        if (student.getClassRoom().getCourse().getDepartment().getId() == department.getId()) {
+            if (!student.isVerified()) {
+                student.getClassRoom().getStudents().remove(student);
+                student.setClassRoom(null);
+                student.getSubjects().stream()
+                        .forEach(e -> e.getStudents().remove(student));
+                student.getSubjects().clear();
+                student.getAttendance().stream()
+                        .forEach(e -> e.setStudent(null));
+                student.getAttendance().clear();
+                session.createCriteria(Attendance.class)
+                        .add(Restrictions.isNull("student"))
+                        .list()
+                        .stream()
+                        .forEach(e -> session.delete(e));
+                student.unaccount();
+            }
+            resp.sendRedirect("/OAS/teacher/hod/students/detailstudent?studentId=" + studentId);
+        } else {
+            resp.sendRedirect("/OAS/error");
         }
-        resp.sendRedirect("/OAS/teacher/hod/students/detailstudent?studentId="+studentId);
     }
-
+    
 }
